Simplify the generic load/save methods
Ok, so, I know what was wrong here. With a method signature like this:

    pub fn load<T: io::Read>(reader: &mut T)

I had to pass &mut &[u8] to load in order for the thing to work
correctly, because io::Read is implemented for &[u8] and I was adding
&mut on top of that in the reader parameter type – unnecessarily. Once I
removed that everything has fallen into place.

In the most problematic place once I removed &mut from &mut &*data what
was left was &*data and that's just converting a Vec to a reference to
a slice and this can be done more ergonomically.