Implement loading base images from graf.dat
This is not yet splitting the images into logical pieces like mouse
cursors, buttons etc. – this will come later as there's quite a bit of
code involved. First things first.

The code can be tested with the following commands:

    cargo run --bin openpol-grafdat-to-image13h /PATH/TO/GRAF.DAT > grafdat13h
    cargo run --bin openpol-image13h-to-ppm -- grafdat13h /PATH/TO/PAL.DAT PALETTE_INDEX > output.ppm
    convert output.ppm output.png
    open output.png

A dummy graf.dat file (dummy_graf.dat) is provided where each of the 15
images is filled with color corresponding to the images index (0, 1, 2,
..., 14) to help with verifying loading and saving the graf.dat contents.
The file is expected to be modified in the future to facilitate testing
splitting the base images into logical pieces.