Eliminate startup/teardown audio clicks
Those were related to the fact that I was filling the buffer 0s when
I wanted silence, while it's supposed to be 128s (with the channel type
u8 that is). Since SDL provides us with the actual value that represents
silence in the given format let's just use it.

There are some clicks remaining when intro audio starts playing or
switches to the next one, but this may be data issue, I'll investigate.