Update dependencies
I mainly need this, because for some reason I couldn't dbg!()
sdl2::event::Event enum before the update. That is – I could, it's just
that it only printed the name of the enum, not the contents. Now it
prints everything so I'm happy.

Also rust-sdl2 0.33 fixes building SDL2 from source in combination with
static linking into the resulting binary, so I may try that soon as
well.