Read the inputs in the main menu
There was no way to know what key or which mouse button was pressed so
there was some refactoring to do. Additionally we now have a Rect method
to test if a given point lies in it.

No actions executed on main menu key or mouse button presses yet, only
placeholders.